==> 01_orte_finden.ipynb <==
### Finden aller potenzieller Orte für die Analyse

- liest die CSV mit allen Wetterstationen ein
- erstellt ein subset mit allen Wetterstationen für die Daten der letzten 10 Jahre vorliegen
- erzeugt daraus die Datei stations.txt 

# Liste mit allen Wetterstationen laden
orte <- read.table("datasets/KL_Tageswerte_Beschreibung_Stationen_UTF8.csv", header = TRUE, sep=",", fill = FALSE, fileEncoding = "UTF-8")
orte$von_datum <- as.Date(as.character(orte$von_datum), format = "%Y%m%d")
orte$bis_datum <- as.Date(as.character(orte$bis_datum), format = "%Y%m%d")

# Anzahl der Orte ausgeben, Liste der Wetterstationen anzeigen und zusammenfassende Statistik
print(paste("Anzahl aller Orte:", nrow(orte)))
head(orte)
summary(orte)

# Zeitraum definieren
start_datum <- as.Date("2014-01-01")
ende_datum <- as.Date("2023-12-31")

# Subset der Wetterstationen für den 10 Jahres Zeitraum erstellen
orte_subset <- subset(orte, von_datum <= ende_datum & bis_datum >= start_datum)

# Anzahl der Orte im Zeitraum 2014-2024 ausgeben
head(orte_subset)
print(paste("Anzahl der Orte im Zeitraum 2014-2024:", nrow(orte_subset)))


# Liste der Wetterstationen in eine Textdatei schreiben und speichern
station_names <- orte_subset$Stationsname
writeLines(station_names, "datasets/stations.txt")
