==> 02_orte_aussuchen.ipynb <==
### Zehn Orte anhand Heuristiken aussuchen

- Von 01_orte_finden erzeugte Stationsliste laden
- Ortslisten anhand der Heuristiken definieren
- Schnittmenge von Stationslisten und Ortslisten bilden

# Lade die Liste der Orte
file_path <- 'datasets/stations.txt'
places <- readLines(file_path, encoding = "UTF-8")

# Orte, die für ihre Nähe zum Meer, zu Bergen und zu Seen in Deutschland bekannt sind
near_sea <- c("Cuxhaven", "Borkum-Süderstraße", "Norderney", "Helgoland", "Warnemünde", "Bremerhaven", "Flensburg(Schäferhaus)", "Travemünde", "Rostock-Warnemünde", "Westerland(Sylt)", "Emden")
near_mountains <- c("Zugspitze", "Feldberg/Schwarzwald", "Brocken", "Oberstdorf", "Garmisch-Partenkirchen", "Wasserkuppe", "Fichtelberg", "Hohenpeißenberg", "Altenberg", "Schwarzwald", "Bayerischer Wald")
near_lakes <- c("Konstanz", "Müllheim", "Lindau(SWN)", "Schwerin", "Potsdam", "München-Flughafen", "Friedrichshafen-Unterraderach", "Chieming", "Starnberg", "Tegernsee", "Schönau am Königssee")

# Filtern der Orte basierend auf Heuristiken
near_sea_and_places <- intersect(near_sea, places)
near_lakes_and_places <- intersect(near_lakes, places)
near_mountains_and_places <- intersect(near_mountains, places)

# Ausgabe der Ergebnisse
cat("Orte mit Wetterstationen am Meer: ", length(near_sea_and_places), near_sea_and_places, "\n")
cat("Orte mit Wetterstationen an Seen: ", length(near_lakes_and_places), near_lakes_and_places, "\n")
cat("Orte mit Wetterstationen in den Bergen: ", length(near_mountains_and_places), near_mountains_and_places, "\n")


# Laden der Liste mit allen vorhandenen Stationen und ihren Orten
orte <- read.table("datasets/KL_Tageswerte_Beschreibung_Stationen_UTF8.csv", header = TRUE, sep=",", fill = TRUE, fileEncoding = "UTF-8")
print(paste("Anzahl aller Orte:", nrow(orte)))

# Gewählte Orte auf Basis der vorherigen Ausgabe
gewaehlte_orte <- subset(
    orte, 
    Stationsname %in% c(
        "Cuxhaven",
        "Norderney",
        "Rostock-Warnemünde",
        "Konstanz",
        "Potsdam",
        "Schwerin",
        "Chieming",
        "Hohenpeißenberg",
        "Feldberg/Schwarzwald",
        "Oberstdorf"))

gewaehlte_orte[order(gewaehlte_orte$Stationsname), c("Stations_id", "Stationsname")]
