==> 04_daten_herunterladen.ipynb <==
### Herunterladen der Wetterdaten für die zehn definierten Orte

- Downloadfunktion und Ortsliste definieren
- Daten für jeden Ort herunterladen und entpacken
- Den Pfad für die einzelnen Wetterdateien speichern in datei_liste.txt

import requests
import zipfile
import io
import os

def search_for_link(url: str, search_text: str) -> str:
    response = requests.get(url)
    if response.status_code == 200:
        content = response.text
        found_index = content.find(search_text)
        if found_index != -1:
            start_index = content.rfind('"', 0, found_index)
            start_index += 1
            end_index = content.find('"', found_index)
            link = content[start_index:end_index]
            return link
    return "search_text not found"

cdc_historical_index_url = "https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/historical/"
cdc_recent_index_url = "https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/recent/"

orte = [
    ["00856", "Chieming"],
    ["00891", "Cuxhaven"],
    ["01346", "Feldberg/Schwarzwald"],
    ["02290", "Hohenpeißenberg"],
    ["02712", "Konstanz"],
    ["03631", "Norderney"],
    ["03730", "Oberstdorf"],
    ["03987", "Potsdam"],
    ["04271", "Rostock-Warnemünde"],
    ["04625", "Schwerin"]]

# Find an download the historical data for the 10 locations
ortsdaten_paths = []

for ort in orte:
    # Search for the link
    file_name = search_for_link(cdc_historical_index_url, ort[0])
    link = cdc_historical_index_url + file_name

    # Download the files
    response = requests.get(link)
    zip_file = zipfile.ZipFile(io.BytesIO(response.content))

    # Extract the contents of the zip file
    foldername = link.split("/")[-1].split(".")[0]
    path = "datasets/orte/" + foldername
    zip_file.extractall(path)

    # Close the zip file
    zip_file.close()

    ortsdaten_paths.append(path)

# Empty the file before writing to it
with open("datasets/orte/datei_liste.txt", "w") as file:
    file.truncate(0)

# Find the produkt_klima_tag file in every path 
found_paths = []
for path in ortsdaten_paths:
    for root, dirs, files in os.walk(path):
        for file in files:
            if file.startswith("produkt") and len(file) > len("produkt"):
                found_paths.append(os.path.join(root, file))

# Print and save the found paths
for path in found_paths:
    print(path)
    with open("datasets/orte/datei_liste.txt", "a") as file:
        file.write(path + "\n")
