==> 07_daten_regressionsanalyse.ipynb <==
# Daten einlesen
wetterdaten <- readRDS("datasets/rds/wetterdaten.rds")

# Ort für die Analyse festlegen
norderney <- wetterdaten[[6]]

# Zeitindex als unabhängige Variable hinzufügen, weil Datum als unabhängige Variable nicht funktioniert, weil es nicht numerisch ist
norderney$Zeitindex <- as.numeric(norderney$MESS_DATUM - min(norderney$MESS_DATUM))
head(norderney)

# Unnötige Spalten entfernen, damit cor() auf dataframe angewendet werden kann
norderney$MESS_DATUM <- NULL
norderney$STATIONS_ID <- NULL
head(norderney)

# Regressionsanalyse beginnt immer mit Korrelationsanalyse
cor(norderney)
cor.test(norderney$Zeitindex, norderney$TMK, conf.level = 0.95)

# Heatmap für TMK und Zeitindex 

# Pakete installieren und laden
if (!require(RColorBrewer, quietly = TRUE)) {
  install.packages("RColorBrewer")
}
if (!require(MASS, quietly = TRUE)) {
  install.packages("MASS")
}
library(RColorBrewer)
library(MASS)

# Funktion für Heatmap
heatmap <- function(df_to_plot, file_path) {

    # PNG
    png(file_path, width=800, height=300)

    # Farbpallette für Darstellung
    rf <- colorRampPalette(rev(brewer.pal(11, "Spectral")))
    r <- rf(32)

    # Plot erzeugen
    h2 <- hist(df_to_plot[,2], breaks=25, plot=FALSE)
    k <- kde2d(df_to_plot[,1], df_to_plot[,2], n=25)
    y_label <- "TMK"
    x_label <- "Zeitindex"
    cex = 1.5

    par(mar = c(6,6,1,1), cex.axis = cex)
    layout(matrix(c(1,2), 1, 2, byrow = TRUE), widths = c(5,1) , heights = c(1,3))
    image(k, col = r)
    mtext(side = 1, line = 3, x_label, cex = cex)
    mtext(side = 2, line = 3, y_label, cex = cex)

    # Barplot rechts
    par(mar = c(5.2,0,0.2,1))
    barplot(h2$counts, axes=FALSE, xlim=c(0, max(h2$counts)), space=0, col="red", horiz=TRUE)

    dev.off()
}

# TMK Dataframe erzeugen
df_tmk <- data.frame(norderney$Zeitindex, norderney$TMK)

# TMK Heatmap erzeugen
heatmap(df_tmk, "plots/linear_regression/heatmap_TMK.png" )
