==> 09_daten_forecast.ipynb <==
### ARIMA Modell

# Daten einlesen
wetterdaten <- readRDS("datasets/rds/wetterdaten.rds")

# Ort für die Analyse festlegen
norderney <- wetterdaten[[6]]
head(norderney)

# Jahr und Monat mit format extrahieren
start_year <- as.numeric(format(min(norderney$MESS_DATUM), "%Y"))
start_month <- as.numeric(format(min(norderney$MESS_DATUM), "%m"))

# Zeitreihenobjekt aus TMK erstellen
TMK_ts <- ts(norderney$TMK, start=c(start_year, start_month), frequency=365)

# Ein passendes Modell mit auto.arima() finden
if (!require(forecast, quietly = TRUE)) {
  install.packages("forecast")
}
modell_tmk <- auto.arima(TMK_ts, seasonal=TRUE, stepwise=TRUE, approximation=TRUE)
summary(modell_tmk)


# Erstellen der Vorhersage für die nächsten 3 Jahre (1095 Tage)
future_forecast <- forecast(modell_tmk, h=1095)

# Ergebnisse der Vorhersage darstellen
png_name <- "plots/forecast/norderney_TMK.png"
png(png_name, width=800, height=400)
plot(future_forecast,
  type = "l",
  xlab = "Jahr",
  ylab = "TMK",
  main = "")
dev.off()
